NEW: Add the ability to switch between engines [jsdom || cheerio].
Engines available are JSDOM with JQuery and cheerio.
The decision of choice an engine can be made at the top level of scrapinode using the scrapinode.engine property or at creation time of the scraper. Indeed #createScraper(url[,engine],callback) function is now like this, the engine parameter is not mandatory; if the engine parameter is missing, the value of scrapinode.engine is used.