FIX: Remove the argument "next" from the method signature "operator" pass in argument in scrapinode.use(path,content,operator).
the next argument was a function, it was looking like a callback and it was not a callback but was more like a passthrough operator, this could have lead to problems and uncomprehension.
This feature is now replaced, now if you want to passthrough and so reach another "operator" method, you should return "null" and not call the argument "next" which is now unavailable.